fix: fix for tables' styles in studio preview of html xblock
In studio preview of the xblocks there are global styles that remove
borders from elements, including tables and elements that are parts of
the tables.

These global styles would overwrite the browser default styles for table
borders, especially when `border` attribute was set on the table.

The fix provides styles that are loaded only in the studio preview. The
new styles try to fix the issue by overwriting the global styles and
setting the border to the same styles as they would be by default.